ジャンプ攻撃の追加とボールが弾を撃つ処理の途中

+++ ゲームプログラミング実習2021_vs2019/GameProgramming/CXPlayer.h
@@ -29,7 +29,7 @@ public:
 	//ジャンプフラグ
 	int mJump;
 	//ジャンプ用速度
-	int mVj;
+	float mVj;
 };
 
 #endif
